# public/dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TUTO",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://api.tuto.fr/v1",
            "description": "mon api"
        }
    ],
    "paths": {
        "/city/list": {
            "get": {
                "tags": [
                    "City"
                ],
                "operationId": "App\\Controller\\CityController::listCity",
                "responses": {
                    "200": {
                        "description": "Nos Villes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/City"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/city/show/{id}": {
            "get": {
                "tags": [
                    "City"
                ],
                "operationId": "App\\Controller\\CityController::showCity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show City",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitySingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/city/create": {
            "post": {
                "tags": [
                    "City"
                ],
                "operationId": "App\\Controller\\CityController::createCity",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUpdateCity"
                },
                "responses": {
                    "200": {
                        "description": "Create City",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitySingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/city/update/{id}": {
            "put": {
                "tags": [
                    "City"
                ],
                "operationId": "App\\Controller\\CityController::updateCity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUpdateCity"
                },
                "responses": {
                    "200": {
                        "description": "Update City",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitySingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/city/delete/{id}": {
            "delete": {
                "tags": [
                    "City"
                ],
                "operationId": "App\\Controller\\CityController::deleteCity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete City",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitySingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/glassdump/list": {
            "get": {
                "tags": [
                    "GlassDump"
                ],
                "operationId": "App\\Controller\\GlassDumpController::listGlassDump",
                "responses": {
                    "200": {
                        "description": "This GlassDump",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GlassDumpList"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/glassdump/show/{id}": {
            "get": {
                "tags": [
                    "GlassDump"
                ],
                "operationId": "App\\Controller\\GlassDumpController::showGlassDump",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GlassDump",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GlassDumpSingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/glassdump/create": {
            "post": {
                "tags": [
                    "GlassDump"
                ],
                "operationId": "App\\Controller\\GlassDumpController::createGlassDump",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUpdateGlassDump"
                },
                "responses": {
                    "200": {
                        "description": "GlassDump create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GlassDumpSingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/glassdump/update/{id}": {
            "put": {
                "tags": [
                    "GlassDump"
                ],
                "operationId": "App\\Controller\\GlassDumpController::updateGlassDump",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUpdateGlassDump"
                },
                "responses": {
                    "200": {
                        "description": "Update Reussi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GlassDumpSingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/glassdump/delete/{id}": {
            "delete": {
                "tags": [
                    "GlassDump"
                ],
                "operationId": "App\\Controller\\GlassDumpController::deleteGlassDump",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Reussi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GlassDumpSingle"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/NoAuthorized"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "City": {
                "description": "Schema City Général",
                "properties": {
                    "id": {
                        "description": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "county_code": {
                        "description": "departement",
                        "type": "integer"
                    },
                    "region": {
                        "description": "name region",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "date création",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CitySingle": {
                "description": "Listing d'une villle",
                "properties": {
                    "mail_city": {
                        "description": "mail client",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "date de modif",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/City"
                    }
                ]
            },
            "GlassDump": {
                "description": "Schema GlassDump Général",
                "properties": {
                    "id": {
                        "description": "uuid",
                        "type": "string"
                    },
                    "number_borne": {
                        "description": "numéro de la borne",
                        "type": "integer"
                    },
                    "volume": {
                        "description": "Volume de la Benne total",
                        "type": "integer"
                    },
                    "landmark": {
                        "description": "point de repère",
                        "type": "string"
                    },
                    "collect_day": {
                        "description": "Jour de collecte",
                        "type": "string"
                    },
                    "coordonate": {
                        "description": "geography de postgis",
                        "type": "number",
                        "format": "double"
                    },
                    "created_at": {
                        "description": "date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "date de de modif",
                        "type": "string",
                        "format": "date-time"
                    },
                    "dammage": {
                        "description": "Endommagé ou non",
                        "type": "boolean"
                    },
                    "is_full": {
                        "description": "Pleine ou non",
                        "type": "boolean"
                    },
                    "is_enable": {
                        "description": "Active ou non",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "GlassDumpList": {
                "description": "Schema de lit GlassDump",
                "properties": {
                    "number_borne": {
                        "description": "numéro de la borne",
                        "type": "integer"
                    },
                    "coordonate": {
                        "description": "geography de postgis",
                        "type": "number",
                        "format": "double"
                    },
                    "created_at": {
                        "description": "date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "dammage": {
                        "description": "Endommagé ou non",
                        "type": "boolean"
                    },
                    "is_full": {
                        "description": "Pleine ou non",
                        "type": "boolean"
                    },
                    "is_enable": {
                        "description": "Active ou non",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "GlassDumpSingle": {
                "description": "Only GlassDump",
                "properties": {
                    "volume": {
                        "description": "Volume de la Benne total",
                        "type": "integer"
                    },
                    "landmark": {
                        "description": "point de repère",
                        "type": "string"
                    },
                    "collect_day": {
                        "description": "Jour de collecte",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "date de de modif",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/GlassDumpList"
                    }
                ]
            }
        },
        "responses": {
            "NotFound": {
                "description": "La ressource n'existe pas",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Cette ressource n'existe pas"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NoAuthorized": {
                "description": "Vous n'avez pas accès à cette page",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Vous n'avez pas l'autorisation"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "id de la ressource",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "requestBodies": {
            "CreateUpdateCity": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "county_code",
                                "region"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "county_code": {
                                    "type": "integer"
                                },
                                "region": {
                                    "type": "string"
                                },
                                "mail_city": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CreateUpdateGlassDump": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "number_one",
                                "volume",
                                "landmark",
                                "collect_day",
                                "coordonate"
                            ],
                            "properties": {
                                "number_one": {
                                    "type": "integer"
                                },
                                "volume": {
                                    "type": "integer"
                                },
                                "landmark": {
                                    "type": "string"
                                },
                                "collect_day": {
                                    "type": "string"
                                },
                                "coordonate": {
                                    "description": "geography PostGis",
                                    "type": "number",
                                    "format": "double"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}